streamlit_requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
supabase==2.3.4
python-dotenv==1.0.0
